+++ extract_msg/properties.py
@@ -5,7 +5,8 @@ import pprint
 from . import constants
 from .enums import Intelligence, PropertiesType
 from .prop import createProp
-from .utils import divide, fromTimeStamp, filetimeToUtc, properHex
+from .utils import divide, properHex
+
 
 logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())